Refactor root component
